requirements.txt
django-ninja
djangorestframework-simplejwt
django-environ
psycopg2-binary
gunicorn
coverage
django-ninja-jwt
python-dotenv
Django
django-cors-headers